Fix some bugs & add new code to /routes/inventory
 - Thumbnail can be left out for POST and PUT requests

 - Complete the controller for PUT --> /inventory/:toyID

+++ routes/inventory.js
@@ -5,24 +5,30 @@ var router = require("express").Router({ mergeParams: true }),
 
 router.get("/", (_req, res, next) => {
     Toy.find({}).lean()
-    .then((foundToys) => res.render("management", { route: "inventory", toys: foundToys}))
-    .catch(next)
+        .then((foundToys) => res.render("management", { route: "inventory", toys: foundToys }))
+        .catch(next)
 })
 
 router.post("/", populate.any(), (req, res, next) => {
     var toy = req.body.toy;
-    toy.thumbnail = "/images/uploaded/" + req.files[0].filename;
+    if (req.files.length === 1) toy.thumbnail = "/images/uploaded/" + req.files[0].filename;
     Toy.create(toy)
         .then((newToy) => res.status(201).json(newToy))
         .catch((err) => {
-            fs.unlink("./" + req.files[0].path, () => { }) // Delete the saved image of the error item
+            if (req.files.length === 1) fs.unlink("./" + req.files[0].path, () => { }) // Delete the saved image of the error item
             next(err)
         })
 })
 
 router.put("/:toyID", populate.any(), (req, res, next) => {
-    console.log(req.body)
-    console.log(req.params.toyID)
+    var toy = req.body.toy
+    if (req.files.length === 1) toy.thumbnail = "/images/uploaded/" + req.files[0].filename;
+    Toy.findByIdAndUpdate(req.params.toyID, toy, { new: true, omitUndefined: true }).lean()
+        .then((updated) => res.status(201).json(updated))
+        .catch((err) => {
+            if (req.files.length === 1) fs.unlink("./" + req.files[0].path, () => { }) // Delete the saved image of the error item
+            next(err)
+        })
 })
 
 router.delete("/:toyID", (req, res, next) => {
